+++ Facade/UserManager.cs
@@ -3,6 +3,9 @@
     using BusinessObjects;
     using DataObjects;
     using Microsoft.AspNet.Identity;
+    using System;
+    using System.Data.Entity;
+    using System.Linq;
     using System.Threading.Tasks;
 
     public class UserManager : UserManager<User, int>
@@ -17,6 +20,13 @@
         internal new UserStore Store { get { return (UserStore)base.Store; } }
 
         public string ConfirmEmailUrl { get; set; }
+
+        public async Task<UserClaim> AddTempTokenClaimAsync(User user)
+        {
+            var claim = Store.AddTempTokenClaim(user);
+            await Store.SaveChangesAsync();
+            return claim;
+        }
 
         public async Task<IdentityResult> ChangeEmailAsync(int userId, string email)
         {
@@ -72,26 +82,96 @@
                 // Send confirmation email
                 await SendConfirmationEmailAsync(user.Id);
             }
-            
+
             return result;
         }
 
-        public async Task DeleteUserResourcePool(int resourcePoolId)
+        public async Task<IdentityResult> CreateAsync(User user, UserLoginInfo userLoginInfo)
         {
-            await Store.DeleteUserResourcePool(resourcePoolId);
+            // Email confirmed
+            user.EmailConfirmed = true;
+
+            // Temp token: Since this is an external login, create temp token; it's going to be used to retrieve the real token by the client
+            Store.AddTempTokenClaim(user);
+
+            var result = await base.CreateAsync(user);
+
+            if (result.Succeeded)
+            {
+                await Store.AddLoginAsync(user, userLoginInfo);
+                await Store.SaveChangesAsync();
+            }
+
+            return result;
+        }
+
+        public async Task DeleteUserResourcePoolAsync(int resourcePoolId)
+        {
+            await Store.DeleteUserResourcePoolAsync(resourcePoolId);
             await Store.SaveChangesAsync();
         }
 
-        public async Task DeleteUserElementField(int elementFieldId)
+        public async Task DeleteUserElementFieldAsync(int elementFieldId)
         {
-            await Store.DeleteUserElementField(elementFieldId);
+            await Store.DeleteUserElementFieldAsync(elementFieldId);
             await Store.SaveChangesAsync();
         }
 
-        public async Task DeleteUserElementCell(int elementCellId)
+        public async Task DeleteUserElementCellAsync(int elementCellId)
         {
-            await Store.DeleteUserElementCell(elementCellId);
+            await Store.DeleteUserElementCellAsync(elementCellId);
             await Store.SaveChangesAsync();
+        }
+
+        public async Task<User> FindByTempToken(string tempToken)
+        {
+            // Search for the user
+            var entity = await Users.Include(user => user.Claims).SingleOrDefaultAsync(user => user.Claims.Any(claim => claim.ClaimType == "TempToken" && claim.ClaimValue == tempToken));
+
+            // Return null if there is no..
+            if (entity == null)
+                return null;
+
+            // Delete temp token
+            var tempTokenClaim = entity.Claims.Single(claim => claim.ClaimType == "TempToken");
+            await Store.DeleteUserClaimAsync(tempTokenClaim.Id);
+            await Store.SaveChangesAsync();
+
+            // Return the user
+            return entity;
+        }
+
+        /// <summary>
+        /// For testing purposes
+        /// </summary>
+        /// <returns></returns>
+        public string GetUniqueEmail()
+        {
+            var year = DateTime.Now.Year;
+            var month = DateTime.Now.Month;
+            var day = DateTime.Now.Day;
+            var hour = DateTime.Now.Hour;
+            var minute = DateTime.Now.Minute;
+            var second = DateTime.Now.Second;
+            return "local_" + year + month + day + "_" + hour + minute + second + "@forcrowd.org";
+        }
+
+        public async Task<IdentityResult> LinkLoginAsync(User user, UserLoginInfo userLoginInfo)
+        {
+            // Email confirmed
+            user.EmailConfirmed = true;
+
+            // Temp token: Since this is an external login, create temp token; it's going to be used to retrieve the real token by the client
+            Store.AddTempTokenClaim(user);
+
+            var result = await base.AddLoginAsync(user.Id, userLoginInfo);
+
+            if (result.Succeeded)
+            {
+                await Store.SaveChangesAsync();
+            }
+
+            return result;
         }
 
         public async Task SendConfirmationEmailAsync(int userId, bool resend = false)
@@ -99,14 +179,17 @@
             var subject = "Confirm your email";
             if (resend) subject += " - Resend";
 
+            var user = await base.FindByIdAsync(userId);
             var token = await base.GenerateEmailConfirmationTokenAsync(userId);
             var encodedToken = System.Net.WebUtility.UrlEncode(token);
 
-            var confirmEmailUrl = string.Format("{0}/?token={1}", ConfirmEmailUrl, encodedToken);
+            var confirmEmailUrl = string.Format("{0}?token={1}", ConfirmEmailUrl, encodedToken);
 
             var sbBody = new System.Text.StringBuilder();
             sbBody.AppendLine("    <p>");
             sbBody.AppendLine("        <b>Wealth Economy - Confirm Your Email</b><br />");
+            sbBody.AppendLine("        <br />");
+            sbBody.AppendFormat("        Email: {0}<br />", user.Email);
             sbBody.AppendLine("        <br />");
             sbBody.AppendLine("        Please click the following link to confirm your email address<br />");
             sbBody.AppendFormat("        <a href='{0}'>Confirm your email address</a>", confirmEmailUrl);