+++ Framework/AppSettings.cs
@@ -13,11 +13,67 @@ namespace forCrowd.WealthEconomy.Framework
         }
 
         /// <summary>
+        /// Registration emails will be send from this address
+        /// </summary>
+        public static string RegistrationEmailAddress
+        {
+            get { return ConfigurationManager.AppSettings["RegistrationEmailAddress"]; }
+        }
+
+        /// <summary>
         /// Alert emails will be send to this address
         /// </summary>
         public static string AlertEmailAddress
         {
             get { return ConfigurationManager.AppSettings["AlertEmailAddress"]; }
+        }
+
+        /// <summary>
+        /// Facebook app id
+        /// </summary>
+        public static string FacebookAppId
+        {
+            get { return ConfigurationManager.AppSettings["FacebookAppId"]; }
+        }
+
+        /// <summary>
+        /// Facebook app secret
+        /// </summary>
+        public static string FacebookAppSecret
+        {
+            get { return ConfigurationManager.AppSettings["FacebookAppSecret"]; }
+        }
+
+        /// <summary>
+        /// Google client id
+        /// </summary>
+        public static string GoogleClientId
+        {
+            get { return ConfigurationManager.AppSettings["GoogleClientId"]; }
+        }
+
+        /// <summary>
+        /// Google client secret
+        /// </summary>
+        public static string GoogleClientSecret
+        {
+            get { return ConfigurationManager.AppSettings["GoogleClientSecret"]; }
+        }
+
+        /// <summary>
+        /// Microsoft client id
+        /// </summary>
+        public static string MicrosoftClientId
+        {
+            get { return ConfigurationManager.AppSettings["MicrosoftClientId"]; }
+        }
+
+        /// <summary>
+        /// Microsoft client secret
+        /// </summary>
+        public static string MicrosoftClientSecret
+        {
+            get { return ConfigurationManager.AppSettings["MicrosoftClientSecret"]; }
         }
     }
 }