+++ Web/App/dataContext.js
@@ -194,7 +194,7 @@
                     return promise.then(success).catch(failed).finally(completed);
 
                     function success(result) {
-                        logger.logSuccess('Saved ' + count + ' change(s)', null, false);
+                        logger.logSuccess('Saved ' + count + ' change(s)');
                         return result;
                     }
 
@@ -202,7 +202,7 @@
                         if (typeof error.status !== 'undefined' && error.status === '409') {
                             logger.logError('Save failed!<br />The record you attempted to edit was modified by another user after you got the original value. The edit operation was canceled.', error, false);
                         } else {
-                            logger.logError('Save failed!', error, false);
+                            logger.logError('Save failed!', error);
                         }
                         return $q.reject(error); // pass error along to next handler
                     }