+++ Web/App_Start/FilterConfig.cs
@@ -1,12 +1,14 @@
-﻿using System.Web.Mvc;
-
-namespace forCrowd.WealthEconomy.Web
+﻿namespace forCrowd.WealthEconomy.Web
 {
+    using System.Web.Mvc;
+
     public class FilterConfig
     {
         public static void RegisterGlobalFilters(GlobalFilterCollection filters)
         {
             filters.Add(new HandleErrorAttribute());
+            // TODO This doesn't do anything?
+            filters.Add(new RequireHttpsAttribute());
         }
     }
 }