+++ Web/Filters/RequireHttpsAttribute.cs
@@ -1,0 +1,26 @@
+﻿namespace forCrowd.WealthEconomy.Web.Filters
+{
+    using System;
+    using System.Net;
+    using System.Net.Http;
+    using System.Web.Http.Controllers;
+    using System.Web.Http.Filters;
+
+    public class RequireHttpsAttribute : AuthorizationFilterAttribute
+    {
+        public override void OnAuthorization(HttpActionContext actionContext)
+        {
+            if (actionContext.Request.RequestUri.Scheme != Uri.UriSchemeHttps && !actionContext.RequestContext.IsLocal)
+            {
+                actionContext.Response = new HttpResponseMessage(HttpStatusCode.Forbidden)
+                {
+                    ReasonPhrase = "HTTPS Required"
+                };
+            }
+            else
+            {
+                base.OnAuthorization(actionContext);
+            }
+        }
+    }
+}