+++ Web/Global.asax.cs
@@ -1,11 +1,11 @@
-﻿using System.Web;
-using System.Web.Http;
-using System.Web.Mvc;
-using System.Web.Optimization;
-using System.Web.Routing;
-
-namespace forCrowd.WealthEconomy.Web
+﻿namespace forCrowd.WealthEconomy.Web
 {
+    using System.Web;
+    using System.Web.Http;
+    using System.Web.Mvc;
+    using System.Web.Optimization;
+    using System.Web.Routing;
+
     public class WebApiApplication : HttpApplication
     {
         protected void Application_Start()
@@ -25,6 +25,10 @@ namespace forCrowd.WealthEconomy.Web
 
         protected void Application_BeginRequest(object sender, System.EventArgs e)
         {
+            // TODO Try to handle this with a filter?
+            if (!Context.Request.IsSecureConnection && !Context.Request.IsLocal)
+                Response.Redirect(Context.Request.Url.ToString().Replace("http:", "https:"));
+
             // TODO Try to create or use MVC like routing, then remove MVC forever?
         }
     }