+++ SolutionItems/Documents/Release Notes.md
@@ -1,5 +1,15 @@
 ﻿### Release Notes
 
+**0.41**
+
+* Social logins: Facebook, Google, Microsoft
+** Social Buttons for Bootstrap: http://lipis.github.io/bootstrap-social/
+** FontAwesome: http://fortawesome.github.io/Font-Awesome/
+* Registration email address info moved to appSettings
+* Elmah settings moved to Configs folder
+* RequireHttps attribute for WebApi
+* Redirect http requests to https for live
+
 **0.40**
 
 * Email confirmation has implemented